Remove need for exported memoized types
